requirements.txt
pymupdf
sqlalchemy
pytesseract
Pillow
fastapi
uvicorn[standard]
python-multipart

